Don't redirect immediately after attempting to remove all admins from a group
On the original implementation, when you attempted to remove all admins from a
group via the Dashboard panel, you were immediately redirected away from the
screen, without the rest of the script being processed. This meant, in
particular, that the bp_group_admin_edit_after action was not run, with the
result that plugins weren't able to save their options properly in this
scenario.

This changeset modifies the process so that, in the case of this kind of error,
the redirection doesn't happen until the end of the form handler, allowing
plugins to hook into the save routine correctly.

Fixes #4930

git-svn-id: https://buddypress.svn.wordpress.org/branches/1.7@6913 cdf35c40-ae34-48e0-9cc9-0c9da1808c22