Notifications:
Make `buddypress()->active_components` the default value for the `component_name` argument.

This is necessary because notifications are persistent in the database, and active components register callbacks for generating output based on the queried results. If notifications exist for components that are inactive, no callback will be registered, and the output will be empty. By only querying for active components, we can be sure no empty notifications will be displayed to the user.

Reverts r7677. See #5290. Fixes #5300. (1.9 branch)

git-svn-id: https://buddypress.svn.wordpress.org/branches/1.9@7687 cdf35c40-ae34-48e0-9cc9-0c9da1808c22