Check the right capabilities are used in the BP_Admin class
The introduction of the link "About BuddyPress" within the WP Admin Bar revealed that in multisite configurations, some administration menus were accessible to regular administrators when they should be restricted to the Super Administrator.
Now, the capacity used in these menus will be adapted to the configuration in which BuddyPress evolves using a "network" capacity when necessary. 

props johnjamesjacoby, boonebgorges, imath

See #5465

git-svn-id: https://buddypress.svn.wordpress.org/branches/1.9@8139 cdf35c40-ae34-48e0-9cc9-0c9da1808c22