Check get_current_screen() is available before using it
In non multisite configs, to avoid displaying the signups (user_status set to 2) into the users administration screen, a filter to the user query is applyed after checking the current screen is the users one. It appears to be problematic on some configs. So we need to check the function get_current_screen() exists just before using it.

see #5551

git-svn-id: https://buddypress.svn.wordpress.org/branches/2.0@8303 cdf35c40-ae34-48e0-9cc9-0c9da1808c22