Messages: Updates to `bp_get_the_message_star_action_link()`:
* Set default `$user_id` so links work on displayed users
* Use `switch` vs. `if/else` to make `$user_domain` calculation more obvious
* Bail if `$user_domain` cannot be calculated
* Cast a few variables to avoid non-integer issues cascading into other methods
* Strict boolean comparisons to ensure proper execution
* Escape output of a few HTML attributes

This change allows for the action-link used to star a message to use the correct URL for the `user_id` that's passed into it. This is particularly helpful for capable community moderators that can view other user's Messages, so the links point to the correct URLs for the correct user.

See #6489. (2.3 branch, for 2.3.2)

git-svn-id: https://buddypress.svn.wordpress.org/branches/2.3@9926 cdf35c40-ae34-48e0-9cc9-0c9da1808c22