JS: Fix an issue with parsing the page number when clicking on a pagination link.
Previously, when a paginated link is clicked, we grabbed the page number
from the URL.  This isn't great as the URL is unreliable to parse the page
number (see #6189).

This commit partially reverts r7946 to use the older page number parsing,
but with some logic to properly parse numbers.

See #6745 (2.4-branch).

Anti-props r-a-y.

git-svn-id: https://buddypress.svn.wordpress.org/branches/2.4@10381 cdf35c40-ae34-48e0-9cc9-0c9da1808c22