Improve input handling on a variety of admin screens.
An internal audit revealed some screens which react to form submission where we were not sanitising input as well as we could have been. We've added extra typecasting and existence checks to avoid mishandling causing PHP Notices.

Additionally, we moved a list of hardcoded BuddyPress "pages" (registration and account activation) into a filterable function to allow plugin developers to more easily customise or add to these, which also gives us the benefit of being able to use that function as a sanitisation whitelist, without duplicating that hardcoded list.

Props boonebgorges, DJPaul


git-svn-id: https://buddypress.svn.wordpress.org/branches/2.4@10385 cdf35c40-ae34-48e0-9cc9-0c9da1808c22