Make sure the Activity Component is active before using functions relative to it inside the Blogs component.
r10544 introduced a side effect on the BuddyPress Emails feature when the Blogs component is activated and the Activity one is deactivated.

This changeset removed some `bp_is_active()` checks inside `bp_blogs_delete_activity()` and `bp_blogs_record_activity()` in favor a conditional load of the whole `bp-blogs/bp-blogs-activity.php` file.

Unfortunately, an issue appeared when the Activity component is not active because these functions are used when some WordPress hooks are fired, for instance when a post type is deleted.

For the 2.5 Branch we are bringing back the `bp_is_active()` checks before using the two functions mentioned earlier. For next release, we will carry on moving all the code specific to the activity component inside the `bp-blogs/bp-blogs-activity.php` file.

See #6937 (branch 2.5)

git-svn-id: https://buddypress.svn.wordpress.org/branches/2.5@10631 cdf35c40-ae34-48e0-9cc9-0c9da1808c22