Core: In our class autoloader, check if the component is active before attempting to include it.
In #6853, we introduced class autoloading to help reduce the amount of
memory used by BuddyPress on any given page request.

This caused a problem in some BuddyPress plugins that previously relied on
a `class_exists()` check to determine if a BuddyPress component is active
or not.  For example, `if ( class_exists( 'BP_Group_Extension' ) ) {}`.

In BuddyPress 2.6.0, `class_exists( 'BP_Group_Extension' )` will now return
true all the time by default.  In some instances, this can cause a fatal
error if the Groups component is disabled.

This commit adds a component check before including the class, while
allowing core classes and PHPUnit to load as usual.

Plugin developers should read the following development post for more
details:
https://bpdevel.wordpress.com/2016/06/28/class-autoloading-and-what-this-means-for-plugin-developers/

Fixes #7140 (2.6-branch).

git-svn-id: https://buddypress.svn.wordpress.org/branches/2.6@10918 cdf35c40-ae34-48e0-9cc9-0c9da1808c22