Activity: Fix activity comment deletion when comment ID equals `0`.
In #7394, non-existent activity ID lookups now default to zero, instead of
the ID that was passed. For example, `$a = new BP_Activity_Activity( 99 )`,
`$a->id` now equals `0` instead of `99`.

This is logistically correct; however, this raised a bug with
`bp_activity_delete_comment()` when passing `0` as the comment ID, which
could potentially delete all activity comments in the database.

Props Maniou.

Fixes #7450 (2.8-branch)

git-svn-id: https://buddypress.svn.wordpress.org/branches/2.8@11584 cdf35c40-ae34-48e0-9cc9-0c9da1808c22