Fix admin user cases in bp_groups_user_can_filter().
If a user is an admin, he or she is pretty much allowed to do anything, so the capability has been approved before `bp_groups_user_can_filter()` filters the value. In a few cases, even admins need to satisfy a few other requirements before being allowed to do something, like request membership in a group, where even site admins need to not be members of the group.

Props r-a-y, dcavins.

See #7610.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@11970 cdf35c40-ae34-48e0-9cc9-0c9da1808c22