Groups: Ensure 'last_activity' meta is updated when the Activity component is disabled.
In r11924, the `bp-groups-activity.php` file is loaded only if the Activity
component is disabled.

However, the `groups_update_last_activity()` function was in this file.
This function is needed to set the `last_activity` group meta entry.  To
ensure the `last_activity` group meta entry is always updated, this commit
moves the `groups_update_last_activity()` function over to
`bp-groups-functions.php`.

Anti-props r-a-y.

See #7870 (3.x branch).

git-svn-id: https://buddypress.svn.wordpress.org/branches/3.0@12123 cdf35c40-ae34-48e0-9cc9-0c9da1808c22