Legacy: Fix activity stream's "Load More" button.
In the bp-legacy template pack, the "Load More" button used to be tied to
the `bp-activity-oldestpage` cookie.  In v3.0, we made steps to remove
this cookie, but this broke activity pagination when used via AJAX (see

This commit outright removes dependencies on the `bp-activity-oldestpage`
cookie and uses a local variable to track the current activity pagination
number.

Props boonebgorges.

See #7896 (trunk).

git-svn-id: https://buddypress.svn.wordpress.org/trunk@12182 cdf35c40-ae34-48e0-9cc9-0c9da1808c22