Attachments: Ensure older bp-default themes can still use the Backbone interface.
Since v3.0.0, `bp-nouveau` is the default template pack.  However, for
themes still using a version of the `bp-default` theme, attempting to
upload results in a fatal error due to a Nouveau function not being
available.

To address this, for bp-default themes, we revert to using the more,
portable `bp-legacy` attachment template parts.

Props imath, r-a-y.

Fixes #7905 (3.x branch).

git-svn-id: https://buddypress.svn.wordpress.org/branches/3.0@12187 cdf35c40-ae34-48e0-9cc9-0c9da1808c22