BP Nouveau: use the same control than WP to set the user password
On the BuddyPress registration page and on the general settings user’s page, BP Nouveau is now using the same control WordPress is using to let users choose their password. By default it generates a strong password to make the user aware of the importance of its robustness. The user can still edit it and if the password is too weak, the submit button is disabled until the user activates the checkbox to force the use of a weak password.

Developers, please note the {{{src/bp-templates/bp-nouveau/js/password-verify.js}}} is now deprecated and will be deleted in version 6.0.0. If you were using the {{{bp-nouveau-password-verify}}} script handle as a dependency of your script(s), please stop doing so before the 6.0.0 release.

Fixes #8092

git-svn-id: https://buddypress.svn.wordpress.org/trunk@12397 cdf35c40-ae34-48e0-9cc9-0c9da1808c22