Introduce a new User Interface to manage Group members
Welcome to the first Core feature to use the BP REST API, although this API is not yet included in BuddyPress! For now this feature is completely silent but as soon as the BP REST API has landed into BuddyPress, you will be able to enjoy a Group member management greatly improved.

1. Whether you prefer to manage Group members from the single Group Administration screen into the WordPress back-end or from the Group Manage screen on the front-end, you will now experience a unique interface to promote, demote, remove or ban the Group members.
2. Very popular Group Administrators will love the new search feature to save some time to find a particular member out of their crowded Group!

On a side note, this commit also introduces the {{{bp-api-request.js}}} JavaScript to ease exchanges with the BP REST API.

Props espellcaste

Fixes #8045

git-svn-id: https://buddypress.svn.wordpress.org/trunk@12405 cdf35c40-ae34-48e0-9cc9-0c9da1808c22