BP Embeds: anticipate a deprecated notice to be introduced in WP 5.3.0
WordPress 5.3.0 has moved the {{{WP_oEmbed}}} class we are using into BP Embeds into a new file, deprecating the older one. To be ready for this next major version of WordPress we are now checking for this new file path to eventually use it if it exists and keeping on using the older one for previous version of WordPress if it is not the case.

Fixes #8115

git-svn-id: https://buddypress.svn.wordpress.org/trunk@12414 cdf35c40-ae34-48e0-9cc9-0c9da1808c22