Add the BP Block components & BP Blocks JavaScript development version
The `src/js` directory is containing the development version of the JavaScript files used to create the first BP Block Component and the two first BP Blocks.

The BP AutoCompleter component is this first BP Block Component, it is used by the BP Member Block and the BP Group Block to let site owners easily select the member or the group they wish to feature into their post. This component uses the BP REST API to suggest results for BP single items.

See #8048


git-svn-id: https://buddypress.svn.wordpress.org/trunk@12576 cdf35c40-ae34-48e0-9cc9-0c9da1808c22