Members: When marking a user as a spammer, do not mark sites as spam if the site has more than one administrator.
Previously on a multisite install, we would mark all the spammer's
sites as spam. This is pretty aggressive and could unintentionally
mark legitmiate sites as spam as well.

To address this, we now only mark a site as spam if the spammer is the
sole administrator of the site.

Fixes #8316 (for 6.0 branch).

git-svn-id: https://buddypress.svn.wordpress.org/branches/6.0@12666 cdf35c40-ae34-48e0-9cc9-0c9da1808c22