BP Members: improve our password validation process
We're introducing a new function to validate the member's chosen password: `bp_members_validate_user_password()`.

This function is primarly used to check the password is not empty, and to make sure the password confirmation matches the password. If it's the case, the function will return a `WP_Error` object with no error message. Otherwise this object will contain an error message.

Plugins can now use the `bp_members_validate_user_password` filter to add their own error messages according to a custom validation process. See the last unit tests of this commit for an example of use.

Props devnik, tharsheblows

Fixes #8066


git-svn-id: https://buddypress.svn.wordpress.org/trunk@12679 cdf35c40-ae34-48e0-9cc9-0c9da1808c22