Blogs: use a default "blavatar" when the blog has no site icon
So far when the Blogs component was not supporting the WordPress site icon feature or when the blog had no site icon set, we used to fallback on the Blog Administrator's avatar.

We realized this was a wrong approach, because:
- the site creator may no longer be a member of the specific site, in which case the lowest id will belong to a different (non-admin) user.
- The site admin may have changed in some other way.
- The site creator's account may have been deleted.
- Etc.

When any of these happen, you get the avatar of a random site user.

That's why just like we do for users or groups, we are now using a default blog avatar (`mysterious-blog.png`) as a fallback when the site icon for the blog is not available.

Props boonebgorges, vapvarun, johnjamesjacoby

Fixes #8179


git-svn-id: https://buddypress.svn.wordpress.org/trunk@12772 cdf35c40-ae34-48e0-9cc9-0c9da1808c22