BP_Invitation: Increment date_modified on key actions.
When an invitation is sent or accepted, the `date_modified`
should be updated. However, if a `date_modified` has been
specified in the invitation or request creation function
(as we do in our unit test functions), the `date_modified`
should be set to that, even if notices are sent (now) as
part of the creation process.

Ensure that BP_Invitation::get_query_clauses() handles specified date_modified data updates.

Fixes #8444.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@12873 cdf35c40-ae34-48e0-9cc9-0c9da1808c22