BP Nouveau: improve the way we get the `nonce` in JavaScript main file
Splitting the `data-bp-nonce` attribute to get Ajax buttons nonce is not the right way to do it. There can be situation where the `_wpnonce` URL param is not the last one (eg: the `data-bp-nonce` attribute of the  Accept Group Invite button contains a `redirect_to` param). Using the `bp.Nouveau.getLinkParams()` function is fixing the issue.

Prop meijioro

Fixes #8396



git-svn-id: https://buddypress.svn.wordpress.org/trunk@12890 cdf35c40-ae34-48e0-9cc9-0c9da1808c22