Remove Travis & raise min. WP version in Unit tests GH action
Bye & thanks for your services `Travis-CI.org`, we decided migrating to `.com` was not for us and preferred to switch to GitHub actions instead.

As we're starting merging Widget Blocks from our BP Blocks feature as a plugin to Core, let's update the min. tested WP version to 5.8 to limit the risks of mistakes.

props johnjamesjacoby

Fixes #8501



git-svn-id: https://buddypress.svn.wordpress.org/trunk@12992 cdf35c40-ae34-48e0-9cc9-0c9da1808c22