Make sure BP directory page IDs are not wrongly unset on multisite
In case a post of a different site of the multisite network has the same ID than one of the BP directory page IDs, the function can wrongly return an array in which the corresponding directory page ID is missing.

We need to make sure this function always run into the BP Root Blog context.

Props thomaslhotta

Fixes #8592



git-svn-id: https://buddypress.svn.wordpress.org/trunk@13159 cdf35c40-ae34-48e0-9cc9-0c9da1808c22