Stop using the `&&` operator inside the Group invites JS templates
The BP Nouveau template pack's Invite UI is failing to render the fetched members when the active theme is supporting the WordPress Full Site editing. The reason is the underscore JS library weirdly doesn't recognize the `&&` operator. To avoid this issue, we've edited the Nouveau Ajax actions/ Group Invites JS templates to set new properties making this operator useless.

This commit also improves the user feedback when invites are restricted to mods or admin (or both) and the current user is a regular group member.

See #8474



git-svn-id: https://buddypress.svn.wordpress.org/trunk@13192 cdf35c40-ae34-48e0-9cc9-0c9da1808c22