Improve behavior of bp_init_group_extensions().
In r13101, the behavior of `bp_register_group_extension()`
was changed to make it possible to modify the behavior via
plugins. However, the same code in a new place behaved
differently, only fully registering the last group extension that
was passed to it. This commit stops passing the group
extension classes by reference, which fixes the issue.

See #8558.

git-svn-id: https://buddypress.svn.wordpress.org/trunk@13195 cdf35c40-ae34-48e0-9cc9-0c9da1808c22