BP Nouveau: improve the  `BP_Buttons_Group` class
Make sure filtering `bp_nouveau_get_activity_entry_buttons` to create a 
button for a specific item's loop is not propagated to other items.

As the buttons of the group are cached, we need to eventually ignore some 
cache data if this filter is used to generate button for specific items of 
the loop instead of all items.

Props bouncingsprout 

Fixes #8716



git-svn-id: https://buddypress.svn.wordpress.org/trunk@13338 cdf35c40-ae34-48e0-9cc9-0c9da1808c22