Silently introduce block support for Activity content
Functions and filters bringing support for Block based activity content are disabled by default. We're adding these so that it's easier to progress on building a Block Based activity post form from a feature as a plugin and the BP Attachments BuddyPress Add-on. To allow Activity content to support blocks, we simply need to use this filter `add_filter( 'bp_is_activity_blocks_active', '__return_true' ); `.

This commit also allowes community users to share more than 2 links when these links are coming from the website BuddyPress is activated on.

Closes https://github.com/buddypress/buddypress/pull/38
Fixes #8765



git-svn-id: https://buddypress.svn.wordpress.org/trunk@13370 cdf35c40-ae34-48e0-9cc9-0c9da1808c22