Improve consistency between private message status & notification
A private message can have a status set to read or unread by the user. When the Notifications component is active, a notification is generated each time a message is sent. When the user opens this message, 2 actions are performed:

1. Set the message status as read.
2. Mark the notification about this message as read.

When opening a message, even if the message has a read status, we still need to check if there's an existing notification about it as notifications can be marked unread from the user's notifications screen. If it's the case, then we need to make sure to mark this unread notification as read 🤪.

Props niftythree

Fixes #8778 (branch 10.0)



git-svn-id: https://buddypress.svn.wordpress.org/branches/10.0@13384 cdf35c40-ae34-48e0-9cc9-0c9da1808c22